Use full width headers and refactor linode tabs
