+++ src/linodes/layouts/IndexPage.js
@@ -216,7 +216,7 @@ export class IndexPage extends Component {
     );
 
     return (
-      <div className={`linodes-page ${view}`}>
+      <div className={`container linodes-page ${view}`}>
         <header>
           <div className="mainmenu">
             <Link to="/linodes/create" className="linode-add btn btn-primary pull-right">
