+++ src/linodes/linode/settings/layouts/DisplayPage.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function DisplayPage() {
+  return (
+    <div className="card">
+      TODO
+    </div>
+  );
+}
