+++ src/linodes/linode/settings/layouts/index.js
@@ -1,2 +1,3 @@
 export IndexPage from './IndexPage';
 export AlertsPage from './AlertsPage';
+export DisplayPage from './DisplayPage';
